move ingredient NBT matching to an upper class
