+++ src/main/java/gregtech/api/recipes/CountableIngredient.java
@@ -1,5 +1,7 @@
 package gregtech.api.recipes;
 
+import gregtech.api.recipes.ingredients.NBTMatching.NBTMatcher;
+import gregtech.api.recipes.ingredients.NBTMatching.NBTcondition;
 import gregtech.api.unification.material.Material;
 import gregtech.api.unification.ore.OrePrefix;
 import gregtech.api.unification.stack.UnificationEntry;
@@ -8,7 +10,6 @@ import gregtech.api.util.IngredientHashStrategy;
 import gregtech.common.ConfigHolder;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.nbt.NBTTagCompound;
 import net.minecraftforge.oredict.OreDictionary;
 import net.minecraftforge.oredict.OreIngredient;
 
@@ -21,7 +22,7 @@ public class CountableIngredient {
     private boolean nonConsumable = false;
     private boolean hasNBTMatchingCondition = false;
     private NBTcondition NBTMatchingCondition;
-    private NBTMatcher NBTMatcher;
+    private gregtech.api.recipes.ingredients.NBTMatching.NBTMatcher NBTMatcher;
 
     public CountableIngredient(Ingredient ingredient, int count) {
         this.ingredient = ingredient;
@@ -126,95 +127,4 @@ public class CountableIngredient {
         return "CountableIngredient{" + "ingredient=" + Arrays.toString(ingredient.getMatchingStacks()) + ", count=" + count + '}';
     }
 
-    /**
-     * a class to build a NBT condition, matching a string, then checking if the NBT matches the condition, then returning if the NBT matches
-     */
-    public static class NBTcondition {
-
-        String nbtKey;
-        long value;
-
-        public static NBTcondition ANY;
-
-        public NBTcondition(String nbtKey, long value) {
-            this.nbtKey = nbtKey;
-            this.value = value;
-        }
-
-    }
-
-    public static class NBTMatcher {
-
-        public NBTMatcher(){
-
-        }
-
-        public static NBTMatcher ANY = new NBTMatcher() {
-            @Override
-            public boolean evaluate(ItemStack stack, NBTcondition NBTcondition) {
-                return true;
-            }
-        };
-
-        public static NBTMatcher LESS_THAN = new NBTMatcher() {
-            @Override
-            public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-                return getKeyValue(stack, nbtKey) < value;
-            }
-        };
-
-        public static NBTMatcher LESS_THAN_OR_EQUALS = new NBTMatcher() {
-            @Override
-            public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-                return getKeyValue(stack, nbtKey) <= value;
-            }
-        };
-
-        public static NBTMatcher EQUALS = new NBTMatcher() {
-            @Override
-            public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-                return getKeyValue(stack, nbtKey) == value;
-            }
-        };
-
-        public static NBTMatcher GREATER_THAN_OR_EQUALS = new NBTMatcher() {
-            @Override
-            public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-                return getKeyValue(stack, nbtKey) >= value;
-            }
-        };
-
-        public static NBTMatcher GREATER_THAN = new NBTMatcher() {
-            @Override
-            public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-                return getKeyValue(stack, nbtKey) > value;
-            }
-        };
-
-        public boolean hasKey(ItemStack stack, String nbtKey) {
-            if (stack.hasTagCompound()) {
-                NBTTagCompound nbt = stack.getTagCompound();
-                if (nbt != null) {
-                    return nbt.hasKey(nbtKey);
-                }
-            }
-            return false;
-        }
-
-        public long getKeyValue(ItemStack stack, String nbtKey) {
-            NBTTagCompound nbt = stack.getTagCompound();
-            if (nbt != null) {
-                return nbt.getLong(nbtKey);
-            }
-            return 0;
-        }
-
-        public boolean evaluate(ItemStack stack, NBTcondition NBTcondition) {
-            return hasKey(stack, NBTcondition.nbtKey) && keyValueMatches(stack, NBTcondition.nbtKey, NBTcondition.value);
-        }
-
-        public boolean keyValueMatches(ItemStack stack, String nbtKey, Long value) {
-            return false;
-        }
-    }
 }
