+++ src/main/java/gregtech/api/recipes/RecipeBuilder.java
@@ -7,6 +7,8 @@ import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.multiblock.CleanroomType;
 import gregtech.api.recipes.Recipe.ChanceEntry;
 import gregtech.api.recipes.recipeproperties.CleanroomProperty;
+import gregtech.api.recipes.ingredients.NBTMatching.NBTMatcher;
+import gregtech.api.recipes.ingredients.NBTMatching.NBTcondition;
 import gregtech.api.recipes.recipeproperties.RecipeProperty;
 import gregtech.api.unification.OreDictUnifier;
 import gregtech.api.unification.material.Material;
@@ -203,7 +205,7 @@ public class RecipeBuilder<R extends RecipeBuilder<R>> {
         return input(item, 1);
     }
 
-    public R input(MetaItem<?>.MetaValueItem item, CountableIngredient.NBTMatcher nbtMatcher, CountableIngredient.NBTcondition nbtCondition) {
+    public R input(MetaItem<?>.MetaValueItem item, NBTMatcher nbtMatcher, NBTcondition nbtCondition) {
         return inputs(CountableIngredient.from(item.getStackForm()).setNBTMatchingCondition(nbtCondition, nbtMatcher));
     }
