+++ src/main/java/gregtech/loaders/recipe/BatteryRecipes.java
@@ -1,9 +1,10 @@
 package gregtech.loaders.recipe;
 
 import gregtech.api.metatileentity.multiblock.CleanroomType;
-import gregtech.api.recipes.CountableIngredient;
 import gregtech.api.recipes.ModHandler;
 import gregtech.api.recipes.ingredients.IntCircuitIngredient;
+import gregtech.api.recipes.ingredients.NBTMatching.NBTMatcher;
+import gregtech.api.recipes.ingredients.NBTMatching.NBTcondition;
 import gregtech.api.unification.material.MarkerMaterials.Color;
 import gregtech.api.unification.material.MarkerMaterials.Tier;
 import gregtech.api.unification.stack.UnificationEntry;
@@ -217,24 +218,24 @@ public class BatteryRecipes {
 
 
         // Battery Recycling Recipes
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_CADMIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_LITHIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_SODIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_CADMIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_LITHIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LV_SODIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_LV).buildAndRegister();
 
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_CADMIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_LITHIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_SODIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_CADMIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_LITHIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_MV_SODIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_MV).buildAndRegister();
 
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_CADMIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_LITHIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_SODIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_CADMIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_LITHIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_HV_SODIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_HV).buildAndRegister();
 
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_EV_VANADIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_SMALL_VANADIUM).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_IV_VANADIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_MEDIUM_VANADIUM).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LUV_VANADIUM,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_LARGE_VANADIUM).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_EV_VANADIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_SMALL_VANADIUM).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_IV_VANADIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_MEDIUM_VANADIUM).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_LUV_VANADIUM, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_LARGE_VANADIUM).buildAndRegister();
 
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_ZPM_NAQUADRIA,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_MEDIUM_NAQUADRIA).buildAndRegister();
-        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_UV_NAQUADRIA,CountableIngredient.NBTMatcher.ANY, CountableIngredient.NBTcondition.ANY).output(BATTERY_HULL_LARGE_NAQUADRIA).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_ZPM_NAQUADRIA, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_MEDIUM_NAQUADRIA).buildAndRegister();
+        EXTRACTOR_RECIPES.recipeBuilder().input(BATTERY_UV_NAQUADRIA, NBTMatcher.ANY, NBTcondition.ANY).output(BATTERY_HULL_LARGE_NAQUADRIA).buildAndRegister();
     }
 
     private static void gemBatteries() {
